requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.9
crispy-bootstrap5==0.6
Django==3.2.6
django-crispy-forms==1.13.0
django-nested-admin==3.4.0
django-stdimage==5.3.0
idna==3.3
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-monkey-business==1.0.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
